--- Autonomous/ardetector/CNN/timestamp.py
@@ -1,5 +1,0 @@
-import time
-
-while True:
-    print(time.time())
-    time.sleep(1)
